File cleanup
With just a few exceptions, this commit finishes the task of  cleaning
up files in the main directory.  Here, I've moved all image files and
supporting files to the appropriate subdirectories with the labs they
belong to, with the exception of a few "orphan" image files I put in
their own subdirectory.  Many duplicate files were deleted, along with a
few obviously useless ones.