New lab: Ohm's law, series and parallel resistors
This is a new lab about Ohm's law and calculating equivalent resistances
in series and parallel.  (Actually I've used this lab for a while, but
just now converted it to Latex.)  This lab plays especially well with
the "Introduction to Electric Circuits" lab I just added.  This commit
also tweaks the master.tex file to include a new Latex command
definition so that the two new circuit labs don't both redefine the same
command.