What's with this one freaking .aux file???????
I included **/*.aux in the .gitignore file, which means that all .aux
files should be ignored.  But this ONE file, heat_vap_nit.aux, keeps
showing up as a changed to be pushed.  What's with this??? Why can't
this one file be ignored?  I'll push it up one more time here.