Major revision of plane waves lab
This set of changes follows up on those three commits back (8/20/2015).
Having reorganized the Induction labs, this was a major revision to the
resulting lab focusing on "plane waves"  Significant editing to the part
that uses the falstad simulation.  (Asked more questions about E cross
B, relationship between spherical and plane waves.)  In activity 2, I
updated screengrab for second simulation reflecting new version of the
applet, and focused more on the meaning of plane waves.  In Activity 3,
changed notation to match the axes of the simulation.