Defined Name/Section/Date header as macro in master.tex
Each lab now simply contains a line at the top, \makelabheader, which is
defined in master.tex.  I think I made the change in ALL of the labs,
even including old ones we're not currently using.  I also changed the
definition of the macro in master.tex to include name and lab partners,
instead of name, section, and date.  But if anybody wants to argue with
me and change it back, they only need to change it in the master.tex
file.