Added indexing of color pages
I used the makeidx package, and manually added the tag \index{color
page} to every place where we use a color figure.  Now master.tex
automatically generates a file master.idx which lists all page numbers
with instances of color figures, which for some reason we need to notify
the printer of.