All formatting files for 132 now in 131 folder
Previously, there were two copies of labmanual_formatting_commands.tex
and labmanual.sty, one in the 131 folder and one in the 132.  I always
worked to keep them identical.  Now I made it so the 132 manual simply
reads the files from the 131 folder.  Also I got rid of fancyhdr.sty,
which is a standard package anyway so should be maintained in the latex
implementation; it is for MikTex, anyway.