New front page and title
New, more interesting and more relevant graphics for front page.  Also
changed the title to "Physics For Doing," which sounds like more of a
page-turner than an "investigative module of activity units."  Also
removed references to where authors are now, and tweaked the "abstract"
which wasn't really an abstract anyway.  Moved various files into a
subdirectory as well.