file cleanup
Several labs are "obsolete" in the sense that we haven't used them in a
long time, AND we have direct replacements for them.  I have removed
these from the master.tex file, and created a new file called
obsolete_labs.tex just so we remember we have them.  Also changed some
directory names, to more closely match the titles of the labs,
eliminating numbered directories like "magnetism_1" and "magnetism_2",
which don't make sense if you're not using all of them anyway.