Added directories and dummy files for Mac compatibility
Mac installation of TeX can't create directories when it needs to, to house .aux files. This causes a problem when using \includelab to include labs from other directories. I created all the necessary directories, and put a dummy file in each to make sure that GitHub maintains the directory structure. Need to do the same if you include any labs from other parts of the directory tree (e.g., those with .. in the path).