Spreadsheet updated for spring, 2020 (NOT FINAL)
I have adjusted the columns for the instructors, keeping only the most recent time that each of us has taught the course.  For spring of 2020, I have prefilled columns for Jack, Jerry, and Matt based on what they did the last time they taught it.  (Note: these might not reflect accurately the places where the labs have changed, such as the changes to the Electric Field labs, or the forking of two separate branches between Jerry and Matt for the diffraction and interference labs.