Changes to e/m lab
I folded material from the old Activity 3 into Activity 1, so that all of the parts pertaining to directions of vectors are grouped together.  I rewrote that hole part with more specific figures, so that students can rely on the right hand rule without needing to specify directions of i, j, and k-hat vectors.  Besides this, I made minor changes to wording, numbering of parts, and formatting.