Included new version of kinetic theory lab in spreadsheet
This is a version of the kinetic theory lab that I (Matt) wrote in spring 2020 during COVID, which does not use simulation software.  It has both advantages and disadvantages w.r.t. the original kinetic theory lab.  Maybe someday someone can recombine the two versions to get the best of both worlds.  For now, I include it here as an option in the master.tex files and the spreadsheet.